Sled instead of RocksDB, use demo code
