+++ .idea/omnipaxosDB-with-tui.iml
@@ -5,10 +5,10 @@
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/coordinator/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/omnipaxos_server/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/kv_store/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/coordinator/target" />
       <excludeFolder url="file://$MODULE_DIR$/target" />
-      <excludeFolder url="file://$MODULE_DIR$/omnipaxos_server/target" />
+      <excludeFolder url="file://$MODULE_DIR$/kv_store/target" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
