+++ docker-compose.yml
@@ -1,12 +1,17 @@
 version: "3.9"
 
+x-common-variables: &common-variables
+  RUST_BACKTRACE: 1
+  NODES: "[1, 2, 3, 4, 5]"
+  CARGO_UNSTABLE_SPARSE_REGISTRY: true
+
 services:
   coordinator:
     build: ./coordinator/
     container_name: coordinator
     hostname: net
     environment:
-      RUST_BACKTRACE: 1
+      <<: *common-variables
       PORT_MAPPINGS: "[[8012,8021],[8013,8031],[8014,8041],[8015,8051],[8023,8032],[8024,8042],[8025,8052],[8034,8043],[8035,8053],[8045,8054]]"
       CLIENT_PORTS: "[8001, 8002, 8003, 8004, 8005]"
       PORT_TO_PID_MAPPING: "[[8001,1],[8002,2],[8003,3],[8004,4],[8005,5],[8012,1],[8013,1],[8014,1],[8015,1],[8021,2],[8023,2],[8024,2],[8025,2],[8031,3],[8032,3],[8034,3],[8035,3],[8041,4],[8042,4],[8043,4],[8045,4],[8051,5],[8052,5],[8053,5],[8054,5]]"
@@ -35,72 +40,53 @@ services:
     stdin_open: true
     tty: true
   
-  op1:
-    build: ./omnipaxos_server/
-    container_name: op1
-    hostname: op1
-    environment: 
-      - RUST_BACKTRACE=1
-      - CARGO_UNSTABLE_SPARSE_REGISTRY=true
-      - PID=1
-      - PEERS=[2,3,4,5]
-      - PEER_ADDRS=["net:8012","net:8013","net:8014","net:8015"]
-      - API_ADDR=net:8001
+  s1:
+    build: kv_store/
+    container_name: s1
+    hostname: s1
+    environment:
+      <<: *common-variables
+      PID: 1
     depends_on:
       - coordinator
 
-  op2:
-    build: ./omnipaxos_server/
-    container_name: op2
-    hostname: op2
-    environment: 
-      - RUST_BACKTRACE=1
-      - CARGO_UNSTABLE_SPARSE_REGISTRY=true
-      - PID=2
-      - PEERS=[1,3,4,5]
-      - PEER_ADDRS=["net:8021","net:8023","net:8024","net:8025"]
-      - API_ADDR=net:8002
+  s2:
+    build: kv_store/
+    container_name: s2
+    hostname: s2
+    environment:
+      <<: *common-variables
+      PID: 2
     depends_on:
       - coordinator
 
-  op3:
-    build: ./omnipaxos_server/
-    container_name: op3
-    hostname: op3
-    environment: 
-      - RUST_BACKTRACE=1
-      - CARGO_UNSTABLE_SPARSE_REGISTRY=true
-      - PID=3
-      - PEERS=[1,2,4,5]
-      - PEER_ADDRS=["net:8031","net:8032","net:8034","net:8035"]
-      - API_ADDR=net:8003
+  s3:
+    build: kv_store/
+    container_name: s3
+    hostname: s3
+    environment:
+      <<: *common-variables
+      CARGO_UNSTABLE_SPARSE_REGISTRY: true
+      PID: 3
     depends_on:
       - coordinator
 
-  op4:
-    build: ./omnipaxos_server/
-    container_name: op4
-    hostname: op4
-    environment: 
-      - RUST_BACKTRACE=1
-      - CARGO_UNSTABLE_SPARSE_REGISTRY=true
-      - PID=4
-      - PEERS=[1,2,3,5]
-      - PEER_ADDRS=["net:8041","net:8042","net:8043","net:8045"]
-      - API_ADDR=net:8004
+  s4:
+    build: kv_store/
+    container_name: s4
+    hostname: s4
+    environment:
+      <<: *common-variables
+      PID: 4
     depends_on:
       - coordinator
 
-  op5:
-    build: ./omnipaxos_server/
-    container_name: op5
-    hostname: op5
-    environment: 
-      - RUST_BACKTRACE=1
-      - CARGO_UNSTABLE_SPARSE_REGISTRY=true
-      - PID=5
-      - PEERS=[1,2,3,4]
-      - PEER_ADDRS=["net:8051","net:8052","net:8053","net:8054"]
-      - API_ADDR=net:8005
+  s5:
+    build: kv_store/
+    container_name: s5
+    hostname: s5
+    environment:
+      <<: *common-variables
+      PID: 5
     depends_on:
       - coordinator
