+++ kv_store/Cargo.toml
@@ -6,11 +6,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-# TODO: it's not nice to depend on git instead of a relative path here, but docker didn't like that
-omnipaxos = { git = "https://github.com/haraldng/omnipaxos", features = ["serde", "macros"] }
-omnipaxos_storage = { git = "https://github.com/haraldng/omnipaxos" }
+omnipaxos = { version = "0.2.0", features = ["serde", "macros"] }
+omnipaxos_storage = "0.2.0"
 tokio = { version = "1", features = ["rt-multi-thread", "sync", "macros", "net", "io-util", "time"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
 lazy_static = "1.4"
-rocksdb = "0.21.0"
+sled = "0.34.7"
