--- omnipaxos_server/src/kv.rs
@@ -1,55 +1,0 @@
-use omnipaxos::storage::Snapshot;
-use omnipaxos::macros::Entry;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct KeyValue {
-    pub key: String,
-    pub value: String,
-}
-
-/// Same as in network actor
-#[derive(Debug, Clone, Serialize, Deserialize, Entry)]
-pub enum KVCommand {
-    Put(KeyValue),
-    Delete(String),
-    Get(String),
-}
-
-// impl Entry for KVCommand {
-//     type Snapshot = KVSnapshot;
-// }
-//
-// #[derive(Clone, Debug, Serialize, Deserialize)]
-// pub struct KVSnapshot {
-//     snapshotted: HashMap<String, String>,
-// }
-//
-// impl Snapshot<KVCommand> for KVSnapshot {
-//     fn create(entries: &[KVCommand]) -> Self {
-//         let mut snapshotted = HashMap::new();
-//         for e in entries {
-//             match e {
-//                 KVCommand::Put(KeyValue { key, value }) => {
-//                     snapshotted.insert(key.clone(), value.clone());
-//                 }
-//                 KVCommand::Delete(key) => {
-//                     snapshotted.remove(key);
-//                 }
-//                 KVCommand::Get(_) => (),
-//             }
-//         }
-//         Self { snapshotted }
-//     }
-//
-//     fn merge(&mut self, delta: Self) {
-//         for (k, v) in delta.snapshotted {
-//             self.snapshotted.insert(k, v);
-//         }
-//     }
-//
-//     fn use_snapshots() -> bool {
-//         true
-//     }
-// }
