--- omnipaxos_server/src/util.rs
@@ -1,4 +1,0 @@
-use std::time::Duration;
-
-pub const ELECTION_TIMEOUT: Duration = Duration::from_millis(100);
-pub const OUTGOING_MESSAGE_PERIOD: Duration = Duration::from_millis(3);
